requirements.txt
crewai
crewai-tools
python-dotenv
pydantic
pandas
